Create Neopixel LED strip
In this example we create 3 functions in order to control the LED strip more creatively. We want that our LED strip generate random colours each LEDs, and the colours change randomly every second, shifting their position and changing the last one.  

The functions take the parameter (np) in order to be generic and reusable with other LED strip. 

We introduce the random module to implements random numbers and utime module to control the sleep. 

The first one creates random numbers for the colours, using the 3 RGB channels, randint function return a random integer in the range (0, 255), and the return statement returns the value of the colours. 

The fill_random function fills the pixels randomly. We use len(np) in order to be consistent with our code using a parameter instead of a global variable. 

The shift_color function let us change the position of the colours, shifting them one position back along thestrip, and changing the last one. 

In order to repeat continuously the code, we call the shift_color function in the while True, using wtite() to method to output and utime.sleep(1) to repeat every second. 

This way, it will look like the random colors are moving along the strip.